fix(types): fix incorrect typescript export declaration
This package is a commonjs package where the main function is exported using `module.exports =
createPlugin.bind(null, newrelic)`. The type declarations however state `export default function
createPlugin`, which is incorrect (it would be correct if the main function were exported at
`modules.exports.default = createPlugin.bind(null, newrelic)`. When you're working with commonjs
this isn't so much of an issue, as most compilers will be using es module interop, but when you're
actually using es modules there is a real different between those two exports, and typescript fails
to compile with the current types. This change updates the type declarations included in the package
to reflect the actual exported code.
